Standardize GetGameDir_s arguments
Make buffer size_t instead of unsigned int.

Thanks Hooman for the suggestion.

+++ op2ext.cpp
@@ -13,7 +13,7 @@
 // It is not used in any way, but must exist to prevent Windows loader errors.
 OP2EXT_API int StubExt = 0;
 
-OP2EXT_API bool GetGameDir_s(char* buffer, unsigned int bufferSize)
+OP2EXT_API bool GetGameDir_s(char* buffer, size_t bufferSize)
 {
 	// Adding "\\" to end of directory is required for backward compatibility.
 	std::string gameDirectory = GetGameDirectory() + "\\";
